Restructure project into package
