+++ tradingbot/services/redis_client.py
