+++ tradingbot/utils/logging_config.py
